+++ cmd/cleanup.go
@@ -11,8 +11,6 @@ import (
 var (
 	allFlag        bool
 	quickFlag      bool
-	orphanFlag     bool
-	minicondaFlag  bool
 	micromambaFlag bool
 	daysOption     int
 )
@@ -26,7 +24,7 @@ After cleanup, they will not be available anymore.`,
 		if common.DebugFlag {
 			defer common.Stopwatch("Env cleanup lasted").Report()
 		}
-		err := conda.Cleanup(daysOption, dryFlag, orphanFlag, quickFlag, allFlag, minicondaFlag, micromambaFlag)
+		err := conda.Cleanup(daysOption, dryFlag, quickFlag, allFlag, micromambaFlag)
 		if err != nil {
 			pretty.Exit(1, "Error: %v", err)
 		}
@@ -37,8 +35,6 @@ After cleanup, they will not be available anymore.`,
 func init() {
 	configureCmd.AddCommand(cleanupCmd)
 	cleanupCmd.Flags().BoolVarP(&dryFlag, "dryrun", "d", false, "Don't delete environments, just show what would happen.")
-	cleanupCmd.Flags().BoolVarP(&orphanFlag, "orphans", "o", false, "Cleanup orphan, unreachable enviroments.")
-	cleanupCmd.Flags().BoolVarP(&minicondaFlag, "miniconda", "m", false, "Remove miniconda3 installation (replaced by micromamba).")
 	cleanupCmd.Flags().BoolVarP(&micromambaFlag, "micromamba", "", false, "Remove micromamba installation.")
 	cleanupCmd.Flags().BoolVarP(&allFlag, "all", "", false, "Cleanup all enviroments.")
 	cleanupCmd.Flags().BoolVarP(&quickFlag, "quick", "q", false, "Cleanup most of enviroments, but leave hololib and pkgs cache intact.")